3dgan integration (#118)
* fixed distributed trainer in cyclones use case

* commiting integration of 3dgan scripts

* ADD: Download dataset

* FIX: DDP distributed training with manual optimization

* ADD: log with MLFlow

* Sqaaas code (#88)

* Create sqaaas.yml

* Update sqaaas.yml

* Update sqaaas.yml

* Point to the current repo

* Remove unnecessary checkout step

* Rename step

---------


* Sqaaas code (#89)

* Create sqaaas.yml

* Update sqaaas.yml

* Update sqaaas.yml

* Point to the current repo

* Remove unnecessary checkout step

* Rename step

* ADD: adaptive branch discovery for SQAaaS action

* Update sqaaas.yml

---------


* ADD: draft predictor and saver

* ADD: stub for inference pipeline

* ADD: small docs

* UPDATE: inference pipeline components

* UPDATE: reorg

* ADD: image generation for inference

* update tag

* ADD: threshold

* ADD: draft inference

* ADD: draft inference wf

* ADD: working inference workflow

* ADD: 3D scatter plots

* ADD: Dockerfile + refactor

* ADD: .dockerignore

* Update .dockerignore

* ADD: skip download option

* ADD: cern pipeline.yaml

* UPDATE: dataset loading function

* UPDATE: dataset loading function

* UPDATE conf

* UPDATE refactor

* UPDATE refactor

* UPDATE training docs

* Update readme

* update README

* FIX typo

* Update README

* Update mkdir

* UPDATE data paths

* UPDATE Dockerfile

* UPDATE Dockerfiles

* UPDATE for Singularity execution

* FIX version mismatch

* UPDATE Singularity docs

* Named steps pipe (#100)

* ADD: dict steps pipe

* Relax dependency constraint

* UPDATE Singularity exec command

* UPDATE: Image version

* UPDATE: load components from pipeline

* ADD: docs

* Simplify 3DGAN model config

* ADD: mlflow autologging support for PL trainer

* UPDATE container info

* Refactor

* UPDATE dependencies

* FIX linter problem

* Simplified workflow configuration (#108)

* Add SQAaaS dynamic badge for dev branch (#104)

* Add SQAaaS dynamic badge

* Upgrade to sqaaas-assessment-action@v2

* Add draft example

* UPDATE credits field

* ADD docs

* REFACTOR components and pipeline code

* UPDATE docstring

* UPDATE mnist torch uc

* ADD config file parser draft

* ADD itwinaiCLI and ConfigParser

* ADD docs

* ADD pipeline parser and serializer plus tests

* UPDATE docs

* ADD adapter component and tests (incl parser)

* ADD splitter component, improve pipeline, tests

* UPDATE test

* REMOVE todos

* ADD component tests

* ADD serializer tests

* FIX linter

* ADD basic workflow tutorial

* ADD basic intermediate tutorial

* ADD advanced tutorial

* UPDATE advanced tutorial

* UPDATE use cases

* UPDATE save parameters

* FIX linter

* FIX cyclones use case workflow

---------


* Simplified workflow configuration (#109)

* Add SQAaaS dynamic badge for dev branch (#104)

* Add SQAaaS dynamic badge

* Upgrade to sqaaas-assessment-action@v2

* Add draft example

* UPDATE credits field

* ADD docs

* REFACTOR components and pipeline code

* UPDATE docstring

* UPDATE mnist torch uc

* ADD config file parser draft

* ADD itwinaiCLI and ConfigParser

* ADD docs

* ADD pipeline parser and serializer plus tests

* UPDATE docs

* ADD adapter component and tests (incl parser)

* ADD splitter component, improve pipeline, tests

* UPDATE test

* REMOVE todos

* ADD component tests

* ADD serializer tests

* FIX linter

* ADD basic workflow tutorial

* ADD basic intermediate tutorial

* ADD advanced tutorial

* UPDATE advanced tutorial

* UPDATE use cases

* UPDATE save parameters

* FIX linter

* FIX cyclones use case workflow

* ADD slurm jobscript

* FIX merge error

* FIX components template

---------


* ADD integration tests

* FIX test

* FIX 3dgan inference test

* ADD GPU support and update tag

* FIX linter

* ADD override example

* UPDATE 3DGAN inference

* UPDATE inference execution tutorials

* UPDATE README

* UPDATE saver saving sparse tensors

* ADD interlink pods

* UPDATE pod name

* UPDATE annotations

* FIX README

* CLEANUP

* Merge

* update

* ADD tf cpu env

* U[date Makefile

* FIX 3DGAN tests

* FIX data folder path

---------

+++ use-cases/3dgan/saver.py
@@ -1,6 +1,8 @@
 from typing import Dict
 import os
 import shutil
+import pickle
+import random
 
 import torch
 from torch import Tensor
@@ -15,11 +17,13 @@ class ParticleImagesSaver(Saver):
 
     def __init__(
         self,
-        save_dir: str = '3dgan-generated'
+        save_dir: str = '3dgan-generated',
+        aggregate_predictions: bool = False
     ) -> None:
         self.save_parameters(**self.locals2params(locals()))
         super().__init__()
         self.save_dir = save_dir
+        self.aggregate_predictions = aggregate_predictions
 
     @monitor_exec
     def execute(self, generated_images: Dict[str, Tensor]) -> None:
@@ -29,15 +33,31 @@ class ParticleImagesSaver(Saver):
             generated_images (Dict[str, Tensor]): maps unique item ID to
                 the generated image.
         """
-        if os.path.exists(self.save_dir):
-            shutil.rmtree(self.save_dir)
-        os.makedirs(self.save_dir)
+        if self.aggregate_predictions:
+            os.makedirs(self.save_dir, exist_ok=True)
+            sparse_generated_images = dict()
+            for name, res in generated_images.items():
+                sparse_generated_images[name] = res.to_sparse()
+            del generated_images
+            with open(self._random_file(), 'wb') as fp:
+                pickle.dump(sparse_generated_images, fp)
+        else:
+            if os.path.exists(self.save_dir):
+                shutil.rmtree(self.save_dir)
+            os.makedirs(self.save_dir)
+            # Save as torch tensor and jpg image
+            for img_id, img in generated_images.items():
+                img_path = os.path.join(self.save_dir, img_id)
+                torch.save(img, img_path + '.pth')
+                self._save_image(img, img_id, img_path + '.jpg')
 
-        # Save as torch tensor and jpg image
-        for img_id, img in generated_images.items():
-            img_path = os.path.join(self.save_dir, img_id)
-            torch.save(img, img_path + '.pth')
-            self._save_image(img, img_id, img_path + '.jpg')
+    def _random_file(self, extension: str = 'pkl') -> str:
+        fname = "%032x.%s" % (random.getrandbits(128), extension)
+        fpath = os.path.join(self.save_dir, fname)
+        while os.path.exists(fpath):
+            fname = "%032x.%s" % (random.getrandbits(128), extension)
+            fpath = os.path.join(self.save_dir, fname)
+        return fpath
 
     def _save_image(
         self,
